+++ app/controllers/return-logs-setup.controller.js
@@ -10,12 +10,12 @@ const MeterDetailsService = require('../services/return-logs/setup/meter-details
 const MeterProvidedService = require('../services/return-logs/setup/meter-provided.service.js')
 const ReceivedService = require('../services/return-logs/setup/received.service.js')
 const ReportedService = require('../services/return-logs/setup/reported.service.js')
-const StartService = require('../services/return-logs/setup/start.service.js')
+const SubmissionService = require('../services/return-logs/setup/submission.service.js')
 const SubmitMeterDetailsService = require('../services/return-logs/setup/submit-meter-details.service.js')
 const SubmitMeterProvidedService = require('../services/return-logs/setup/submit-meter-provided.service.js')
 const SubmitReceivedService = require('../services/return-logs/setup/submit-received.service.js')
 const SubmitReportedService = require('../services/return-logs/setup/submit-reported.service.js')
-const SubmitStartService = require('../services/return-logs/setup/submit-start.service.js')
+const SubmitSubmissionService = require('../services/return-logs/setup/submit-submission.service.js')
 const SubmitUnitsService = require('../services/return-logs/setup/submit-units.service.js')
 const UnitsService = require('../services/return-logs/setup/units.service.js')
 
@@ -54,11 +54,11 @@ async function setup(request, h) {
   return h.redirect(`/system/return-logs/setup/${session.id}/received`)
 }
 
-async function start(request, h) {
+async function submission(request, h) {
   const { sessionId } = request.params
-  const pageData = await StartService.go(sessionId)
+  const pageData = await SubmissionService.go(sessionId)
 
-  return h.view('return-logs/setup/start.njk', pageData)
+  return h.view('return-logs/setup/submission.njk', pageData)
 }
 
 async function submitMeterDetails(request, h) {
@@ -107,7 +107,7 @@ async function submitReceived(request, h) {
     return h.view('return-logs/setup/received.njk', pageData)
   }
 
-  return h.redirect(`/system/return-logs/setup/${sessionId}/start`)
+  return h.redirect(`/system/return-logs/setup/${sessionId}/submission`)
 }
 
 async function submitReported(request, h) {
@@ -125,13 +125,13 @@ async function submitReported(request, h) {
   return h.redirect(`/system/return-logs/setup/${sessionId}/units`)
 }
 
-async function submitStart(request, h) {
+async function submitSubmission(request, h) {
   const { sessionId } = request.params
 
-  const pageData = await SubmitStartService.go(sessionId, request.payload)
+  const pageData = await SubmitSubmissionService.go(sessionId, request.payload)
 
   if (pageData.error) {
-    return h.view('return-logs/setup/start.njk', pageData)
+    return h.view('return-logs/setup/submission.njk', pageData)
   }
 
   return h.redirect(`/system/return-logs/setup/${sessionId}/reported`)
@@ -165,12 +165,12 @@ module.exports = {
   received,
   reported,
   setup,
-  start,
+  submission,
   submitMeterDetails,
   submitMeterProvided,
   submitReceived,
   submitReported,
-  submitStart,
+  submitSubmission,
   submitUnits,
   units
 }
