+++ app/services/return-logs/setup/submission.service.js
@@ -1,15 +1,15 @@
 'use strict'
 
 /**
- * Orchestrates fetching and presenting the data needed for the `/return-logs/setup/{sessionId}/start` page
- * @module StartService
+ * Orchestrates fetching and presenting the data needed for the `/return-logs/setup/{sessionId}/submission` page
+ * @module SubmissionService
  */
 
 const SessionModel = require('../../../models/session.model.js')
-const StartPresenter = require('../../../presenters/return-logs/setup/start.presenter.js')
+const SubmissionPresenter = require('../../../presenters/return-logs/setup/submission.presenter.js')
 
 /**
- * Orchestrates fetching and presenting the data needed for the `/return-logs/setup/{sessionId}/start` page
+ * Orchestrates fetching and presenting the data needed for the `/return-logs/setup/{sessionId}/submission` page
  *
  * @param {string} sessionId - The UUID of the current session
  *
@@ -18,7 +18,7 @@ const StartPresenter = require('../../../presenters/return-logs/setup/start.pres
 async function go(sessionId) {
   const session = await SessionModel.query().findById(sessionId)
 
-  const formattedData = StartPresenter.go(session)
+  const formattedData = SubmissionPresenter.go(session)
 
   return {
     activeNavBar: 'search',
