--- app/views/return-logs/setup/start.njk
@@ -1,148 +1,0 @@
-{% extends 'layout.njk' %}
-{% from "govuk/components/back-link/macro.njk" import govukBackLink %}
-{% from "govuk/components/button/macro.njk" import govukButton %}
-{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
-{% from "govuk/components/radios/macro.njk" import govukRadios %}
-{% from "govuk/components/summary-list/macro.njk" import govukSummaryList %}
-
-{% from "macros/return-status-tag.njk" import statusTag %}
-
-{% block breadcrumbs %}
-  {# Back link #}
-  {{
-    govukBackLink({
-      text: 'Back',
-      href: backLink
-    })
-  }}
-{% endblock %}
-
-{% block content %}
-  {# Error message #}
-  {% if error %}
-    {{ govukErrorSummary({
-      titleText: "There is a problem",
-      errorList: [
-        {
-          text: error.text,
-          href: '#journey'
-        }
-      ]
-      }) }}
-  {% endif %}
-
-  {# Main heading #}
-  <div class="govuk-grid-row">
-    <div class="govuk-grid-column-full">
-      <span class="govuk-caption-l">Licence {{ licenceRef }}</span>
-      <h1 class="govuk-heading-xl govuk-!-margin-bottom-3">{{ pageTitle }}</h1>
-
-      {# Return status #}
-      <p data-test="return-status">{{ statusTag(status) }}</p>
-
-      {# Return meta-data #}
-      {{ govukSummaryList({
-          classes: 'govuk-summary-list--no-border',
-          rows: [
-            {
-              key: {
-                text: 'Return reference',
-                classes: "meta-data__label"
-              },
-              value: {
-                html: '<span data-test="return-reference">'+ returnReference +'</span>',
-                classes: "meta-data__value"
-              }
-            },
-            {
-              key: {
-                text: 'Site description',
-                classes: "meta-data__label"
-              },
-              value: {
-                html: '<span data-test="site-description">'+ siteDescription +'</span>',
-                classes: "meta-data__value"
-              }
-            },
-            {
-              key: {
-                text: 'Purposes',
-                classes: "meta-data__label"
-              },
-              value: {
-                html: '<span data-test="purposes">'+ purposes +'</span>',
-                classes: "meta-data__value"
-              }
-            },
-            {
-              key: {
-                text: 'Returns period',
-                classes: "meta-data__label"
-              },
-              value: {
-                html: '<span data-test="returns-period">'+ returnsPeriod +'</span>',
-                classes: "meta-data__value"
-              }
-            },
-            {
-              key: {
-                text: 'Abstraction period',
-                classes: "meta-data__label"
-              },
-              value: {
-                html: '<span data-test="abstraction-period">'+ abstractionPeriod +'</span>',
-                classes: "meta-data__value"
-              }
-            },
-            {
-              key: {
-                text: tariffType,
-                classes: "meta-data__label"
-              }
-            }
-          ]
-        }) }}
-    </div>
-  </div>
-
-  {# Select how you would like to edit the return #}
-  <form  method="post">
-    <input type="hidden" name="wrlsCrumb" value="{{wrlsCrumb}}"/>
-    {{ govukRadios({
-      name: "journey",
-      errorMessage: error,
-      fieldset: {
-        legend: {
-          text: "What do you want to do with this return?",
-          isPageHeading: true,
-          classes: "govuk-fieldset__legend--l"
-        }
-      },
-      items: [
-        {
-          value: "enter-return",
-          text: "Enter and submit",
-          checked: journey === "enter-return"
-        },
-        {
-          divider: "or"
-        } if beenReceived === true,
-        {
-          value: "nil-return",
-          text: "Enter a nil return",
-          checked: journey === "nil-return"
-        },
-        {
-          divider: "or"
-        } if beenReceived === false,
-        {
-          value: "record-receipt",
-          text: "Record receipt",
-          checked: journey === "record-receipt"
-        } if beenReceived === false
-      ]
-    }) }}
-
-    {{ govukButton({ text: "Continue", preventDoubleClick: true }) }}
-  </form>
-{% endblock %}
