+++ app/views/return-logs/setup/submission.njk
@@ -1,0 +1,64 @@
+{% extends 'layout.njk' %}
+{% from "govuk/components/back-link/macro.njk" import govukBackLink %}
+{% from "govuk/components/button/macro.njk" import govukButton %}
+{% from "govuk/components/error-summary/macro.njk" import govukErrorSummary %}
+{% from "govuk/components/radios/macro.njk" import govukRadios %}
+
+{% block breadcrumbs %}
+  {# Back link #}
+  {{
+    govukBackLink({
+      text: 'Back',
+      href: backLink
+    })
+  }}
+{% endblock %}
+
+{% block content %}
+  {# Error message #}
+  {% if error %}
+    {{ govukErrorSummary({
+      titleText: "There is a problem",
+      errorList: [
+        {
+          text: error.text,
+          href: '#journey'
+        }
+      ]
+      }) }}
+  {% endif %}
+
+  {# Main heading #}
+  <div>
+    <span class="govuk-caption-l"> Return reference {{ returnReference }} </span>
+    <h1 class="govuk-heading-l">{{ pageTitle }}</h1>
+  </div>
+
+  {# Select what you want to do with this return #}
+  <form  method="post">
+    <input type="hidden" name="wrlsCrumb" value="{{wrlsCrumb}}"/>
+    {{ govukRadios({
+      name: "journey",
+      errorMessage: error,
+      items: [
+        {
+          value: "enter-return",
+          text: "Enter and submit",
+          checked: journey === "enter-return"
+        },
+        {
+          value: "nil-return",
+          text: "Enter a nil return",
+          checked: journey === "nil-return"
+        },
+        {
+          value: "record-receipt",
+          text: "Record receipt",
+          checked: journey === "record-receipt"
+        } if beenReceived === false
+      ]
+    }) }}
+
+    {{ govukButton({ text: "Continue", preventDoubleClick: true }) }}
+  </form>
+{% endblock %}
