+++ test/presenters/return-logs/setup/reported.presenter.test.js
@@ -29,7 +29,7 @@ describe('Return Logs Setup - Reported presenter', () => {
         sessionId: '61e07498-f309-4829-96a9-72084a54996d',
         returnReference: '012345',
         reported: null,
-        backLink: '/system/return-logs/setup/61e07498-f309-4829-96a9-72084a54996d/start'
+        backLink: '/system/return-logs/setup/61e07498-f309-4829-96a9-72084a54996d/submission'
       })
     })
   })
