+++ test/services/return-logs/setup/submit-reported.service.test.js
@@ -56,7 +56,7 @@ describe('Return Logs Setup - Submit Reported service', () => {
             pageTitle: 'How was this return reported?',
             activeNavBar: 'search',
             reported: null,
-            backLink: `/system/return-logs/setup/${session.id}/start`,
+            backLink: `/system/return-logs/setup/${session.id}/submission`,
             returnReference: '12345'
           },
           { skip: ['sessionId', 'error'] }
